+++ dotnet/src/Connectors/Connectors.AI.OpenAI/AzureSdk/TextCompletionStreamingResultBase.cs
@@ -5,24 +5,9 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.AI.OpenAI;
-using Microsoft.SemanticKernel.Connectors.AI.OpenAI.TextCompletion;
+using Microsoft.SemanticKernel.AI.TextCompletion;
 
 namespace Microsoft.SemanticKernel.Connectors.AI.OpenAI.AzureSdk;
-
-internal class TextCompletionResultBase : ITextCompletionResult
-{
-    private readonly Choice _choice;
-
-    public TextCompletionResultBase(Choice choice)
-    {
-        this._choice = choice;
-    }
-
-    public async Task<string> CompleteAsync(CancellationToken cancellationToken = default)
-    {
-        return this._choice.Text;
-    }
-}
 
 internal class TextCompletionStreamingResultBase : ITextCompletionStreamingResult
 {
