+++ dotnet/src/SemanticKernel.Abstractions/AI/TextCompletion/ITextCompletionResult.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.SemanticKernel.AI.TextCompletion;
+
+public interface ITextCompletionResult
+{
+    Task<string> CompleteAsync(CancellationToken cancellationToken = default);
+}
