+++ dotnet/src/SemanticKernel.Abstractions/AI/TextCompletion/ITextCompletionStreamingResult.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+
+using System.Collections.Generic;
+using System.Threading;
+
+namespace Microsoft.SemanticKernel.AI.TextCompletion;
+
+public interface ITextCompletionStreamingResult : ITextCompletionResult
+{
+    IAsyncEnumerable<string> CompleteStreamAsync(CancellationToken cancellationToken = default);
+}
